# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión iniciada."
          },
          "401": {
            "description": "INVALID_CREDENTIALS."
          }
        },
        "summary": "Iniciar sesión con la cuenta opcional",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sesión cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar sesión (siempre 204, con o sin sesión válida)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil del apoderado."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil de la cuenta con sesión activa",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/panel/auth/login": {
      "post": {
        "operationId": "PanelAuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión iniciada."
          },
          "400": {
            "description": "MISSING_FIELDS."
          },
          "401": {
            "description": "INVALID_CREDENTIALS."
          }
        },
        "summary": "Iniciar sesión como educadora",
        "tags": [
          "panel-auth"
        ]
      }
    },
    "/api/panel/auth/logout": {
      "post": {
        "operationId": "PanelAuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sesión cerrada."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar sesión de la educadora (siempre 204)",
        "tags": [
          "panel-auth"
        ]
      }
    },
    "/api/panel/auth/me": {
      "get": {
        "operationId": "PanelAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil de la educadora."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil de la educadora con sesión activa",
        "tags": [
          "panel-auth"
        ]
      }
    },
    "/api/panel/auth/password": {
      "post": {
        "operationId": "PanelAuthController_changePassword",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Clave cambiada; token nuevo."
          },
          "400": {
            "description": "WEAK_PASSWORD."
          },
          "401": {
            "description": "NO_SESSION o INVALID_CREDENTIALS (clave actual incorrecta)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar la clave de la educadora",
        "tags": [
          "panel-auth"
        ]
      }
    },
    "/api/panel/agenda": {
      "get": {
        "operationId": "PanelAgendaController_getWeek",
        "parameters": [
          {
            "name": "weekStart",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Días de la semana con sus cupos."
          },
          "400": {
            "description": "VALIDATION_ERROR: falta weekStart o no es lunes."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agenda semanal (lunes a domingo) de la educadora",
        "tags": [
          "panel-agenda"
        ]
      }
    },
    "/api/panel/summary": {
      "get": {
        "operationId": "PanelDashboardController_get",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen del día pedido (hoy por defecto)."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resumen del día para la educadora",
        "tags": [
          "panel-summary"
        ]
      }
    },
...
    "/api/panel/events": {
      "get": {
        "description": "Emite `event: session` con un `SessionEvent` por cada sesión creada o cambiada, y `event: ping` cada 25 s. No reenvía eventos perdidos: al reconectar, volver a pedir resumen, agenda y notificaciones.",
        "operationId": "PanelEventsController_stream",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Stream abierto (`text/event-stream`)."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream en tiempo real de cambios de sesiones (Server-Sent Events)",
        "tags": [
          "panel-events"
        ]
      }
    },
    "/api/panel/notifications": {
      "get": {
        "operationId": "PanelNotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Hasta 20 avisos recientes y el total de no leídos."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Avisos de la campana: lo que hicieron apoderados o el sistema",
        "tags": [
          "panel-notifications"
        ]
      }
    },
    "/api/panel/notifications/seen": {
      "post": {
        "operationId": "PanelNotificationsController_markSeen",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Avisos marcados como vistos."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marcar todos los avisos como vistos",
        "tags": [
          "panel-notifications"
        ]
      }
    }
  },
  "info": {
    "title": "profesor-scheduling-api",
    "description": "Backend del agendamiento de la educadora diferencial.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}